readme: make it more neutral
I'd like the readme to not look weird in notaz/picodrive when I merge
all the changes, where it will be unclear who the 1st person pronoun
refers to. Several people have put quite some effort into it, so I think
it's best not to highlight anyone.

Also remove the old readme which has become outdated.